build(make): Add build dependencies and update `%-acceptance` targets
The `build` targets tracks the target binary and depends on the source
files. This prevents make from attempting to rebuild the binary when
building the docker image if the binary is already up to date.

The `docker-acceptance` target is turned into a simple dependency on
docker with build flags for adding coverage and other defaults for
docker tag.

The top level `test-acceptance` no longer tears down the compose
environment, but leaves the existing containers running for faster
execution of acceptance tests.

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>