ci: Remove job `publish:frontend:enterprise`
The image will be published from the enterprise fork (for now).
Although there's no enterprise pipeline for the frontend repository, the
publish stage relies on the private CI environment.
The publish job for the enterprise pipeline will do nothing but copy the
image from docker hub to our private repository.

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>