chore(compose): Fix SeaweedFS command passed as single string
There were missing commas in the yaml array, making the whole
s3fs.command a single string. This messes with the entrypoint script in
the SeaweedFS entrypoint which in turn causes some of the backing
services store data outside the data volume (only file index was stored
inside the volume).

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>