refac(compose): Refactor SeaweedFS topology and optimize startup/shutdown time
Instead of running SeaweedFS as a monolith using the `server` command,
we explicitly launch all services in different containers. It appears
that the `server` command has some issues when initializing the master
and sometimes enter a deadlock the healthcheck interval is too low at
startup. Moreover, running the services in different containers makes it
easier to debug and interpret the logs.

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>