ci: drop prepend flag
Git cliff usually work on an existing PR, with a temporary CHANGELOG
with the 4.0.0 version. When another commit for the same version
triggers another git cliff, it runs on the previous pre-populated
CHANGELOG, messing it up. For this reason we're dropping the --prepend
option and always rely on the full git history, which should not change,
to rewrite the full CHANGELOG

Signed-off-by: Roberto Giovanardi <roberto.giovanardi@northern.tech>