feat(deviceconfig): support for the hidden Configure Add-on fields
The set configuration for deviceconfing is extended to accept
fields in the form of an object with `value` and a `hidden` attrbutes

The v2 get configuration for deviceconfig endpoint returns the new
attributes without values, if `hidden` is set to true.

Changelog: It is possible to define a hidden configuration attribute, which will lead to values that will never be openly presented to the user, and can be seen as hidden with queries to the v2 endpoint.
Ticket: MEN-7614
Signed-off-by: Peter Grzybowski <peter@northern.tech>