chore: Use Virtual Device for OS updates in docker composition
For the upcoming release, we will use the same Virtual Device that we
had in the Mender 3.x series. After the release, we will work on
migrating to the Virtual Device for app updates consistently across
docker composition, ui and docs. See more context in the parent JIRA
epic related to this task.

For the e2e tests, use an override image for the `client` service so
that we use the new Virtual Device.

Ticket: MEN-7788

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>