test: client compat migrated to backend integration tests and test separation
Client compatibility test is re-implemented as a backend integration test.

A general mechanism to run tests in a separate and completely independent
session is introduced; all that is needed is to name the test file with
`+separate` before .py extension, for instance moving `test_artifact.py`
to a separate pytest session can be achieved by changing the file name
to `test_artifact+separate.py`.

Ticket: QA-683
Signed-off-by: Peter Grzybowski <peter@northern.tech>