fix: Race when provisioning and connecting a device
If a device gets provisioned and submits a connection request while the
device is getting provisioned, the device might end up in an
inconsistent state where the connection status gets overwritten to
"unknown".
The issue was discovered in a test where the system was under load
and the device was running on the same network (artificially low RTT).

Changelog: Commit
Ticket: MEN-8164
Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>