chore: Add Traefik routing env variable flag: TRAEFIK_LEGACY_INVENTORY
The environment variable will use path replacement to map the public API
to the deprecated internal inventory paths `/api/0.1.0/`.
The flag is disabled by default making use of the newer routing schema.
To run the compose for mender-server <= v4.0 set the environment
variable to a non-empty string or uncomment the variable in the root
.env file.

Ticket: MEN-8216
Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>