fix: list services in Makefile correctly on MacOS
The Makefile function `wildcard`returns both files and directories
in the /services/ directory (at least on Mac). This leads to test
execution failing midway. A simple solution is to keep only directories
using the `dir` function.

In addition, I believe the ordering produced by `wildward` is not
deterministic, so I added a call to `sort` as well to ensure tests are
executed in a consistent order (alphabetical).

Signed-off-by: Frode Hansen <frode.hansen@northern.tech>