fix(deployments): limit `name`, `type` and `device types compatible` length
Limit the max length of `name` (aka `Release name` in the gui),
`type` and `device types compatible` fields in `MultipartGenerateImageMsg`
(used by single file uploads) to 256 characters.

There has been introduced a limit of 256 characters to the corresponding
fields in mender-artifact:
    https://github.com/mendersoftware/mender-artifact/commit/ddd821f8a5150fb8a3186b337525f17b2757599c
causing uploads that do not abide by this limit to fail silently in the
background when being processed by the workflow worker.

With this input validation we provide an explicit error to the user in
this case.

Signed-off-by: Frode Hansen <frode.hansen@northern.tech>