chore(tests): run compose tests on `amd64` always
The integration and acceptance tests use mender-artifact built for x86
(not arm) so these tests always fail if the acceptance-tester and
integration-tester are built for arm64 - which is the case on MacOS
unless a different platform is explicitly defined.

I belive this should make no difference for non MacOS users (CI included)
as this is already how the "testers" are built on those platforms.

Signed-off-by: Frode Hansen <frode.hansen@northern.tech>