fix(inventory): bulk write attributes uses same filter
The bulk update attributes is using the same filter for all updates as
the variable is shared. However, the fix is not externally visible as
the only endpoint that  can trigger the bulk update (Update device
status internal API) is always called with a single device.

Changelog: None
Ticket: MEN-8791
Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>