test: Make ssh_is_opened timeout equal to provided value
Fix ssh_is_opened so that the wait parameter is correctly used, by forwarding it to
MenderDevice::run. Before the fix, the timeout was forwarded to MenderDevice::run,
while the run itself was tried up to "wait" times on top of that.
This resulted in 600*600 seconds (100 hours) of total timeout.
If the client device did not become ready for any reason it caused the pipelines
to seemingly hang.

Changelog: None
Ticket: QA-1147

Signed-off-by: Michal Kopczan <michal.kopczan@northern.tech>