chore: made mockgen.sh work on Mac OS
Updated the `awk` command to escape backslashes so it works with the
version of `awk` shipped with mac os by default as well as `gawk` (which is the
linux default I believe). I tested on a standard ubuntu image and it
seems to work just fine.

Also updated the version of `vektra/mockery` to match enterprise and
make mockgen work with more recent versions of Go.

Ticket: None

Signed-off-by: Frode Hansen <frode.hansen@northern.tech>