feat(workflows): support for accessing data from objects via json
With this change the input parameters that are objects with json
representation can be referenced via `identifier.jsonInput.field` syntax.
For instance: in a device is passed as an input parameter, then
anywhere you use `device.jsonInput.tier` you will refer to the Device.Tier
field, since it is decoreated with json `tier`.

Changelog: Commit
Ticket: MEN-9096
Signed-off-by: Peter Grzybowski <peter@northern.tech>