feat(inventory): internal patch does not override system attributes
With this change the internal PATCH /tenants/:tenant_id/device/:device_id/attribute/scope/:scope
does not override the given attributes with the :scope from url path
if the attribute is of system scope. In other words: what you give in the body
will maintain the scope if it is system scope.

Changelog: Title
Ticket: MEN-9096
Signed-off-by: Peter Grzybowski <peter@northern.tech>